+++ src/ee/liiser/siim/galaxies/data/basic/BasicObject.java
@@ -1,6 +1,6 @@
 package ee.liiser.siim.galaxies.data.basic;
 
-import javax.vecmath.Vector3f;
+import javax.vecmath.Vector3d;
 
 import ee.liiser.siim.galaxies.data.BaseObject;
 
@@ -9,14 +9,14 @@ import ee.liiser.siim.galaxies.data.BaseObject;
  */
 public abstract class BasicObject extends BaseObject{
 	
-	protected Vector3f oldPosition;
+	protected Vector3d oldPosition;
 	
-	public void updatePosition(Vector3f position){
+	public void updatePosition(Vector3d position){
 		oldPosition = this.position;
 		this.position = position;
 	}
 
-	public Vector3f getOldPosition() {
+	public Vector3d getOldPosition() {
 		return oldPosition;
 	}
