+++ src/ee/liiser/siim/galaxies/drawing/Drawable.java
@@ -1,6 +1,6 @@
 package ee.liiser.siim.galaxies.drawing;
 
-import javax.vecmath.Vector3f;
+import javax.vecmath.Vector3d;
 
 /**
  * Interface that wraps all drawable points: All cores and stars
@@ -11,16 +11,16 @@ public interface Drawable {
 	 * 
 	 * @return the position of the object in physical space
 	 */
-	public Vector3f getPosition();
+	public Vector3d getPosition();
 	
 	/**
 	 * Sets the position of the object
 	 * @param position of the object
 	 */
-	public void setPosition(Vector3f position);
+	public void setPosition(Vector3d position);
 
 	/**
 	 * @return Visible size of the object
 	 */
-	public float getSize();
+	public double getSize();
 }
